Declare application.css in manifest for asset precompilation
We encountered an AssetNotPrecompiledError when rendering the layout for the admin components, which revealed that our asset pipeline wasn't including application.scss. This commit adds a declaration in app/assets/config/manifest.js to explicitly link application.scss, ensuring that it gets precompiled. By making this change, our assets are correctly managed in production and the styling for our govuk_publishing_components is reliably loaded without causing layout issues.